rtsp-media: add option indicating pipeline shared between media
By indicating that a media has a shared-pipeline, the media will
not try to manipulate the state of the pipeline nor rely on
pipeline state transitions to prepare itself. If shared-pipeline
is set to TRUE, the MediaFactory is responsible for setting the
pipeline's states. The media will, however, update the state of
its element just as if it were the pipeline. In order to make
it easy for the media to maintain the state of its private
elements (rtpbin, etc.) all elements created by the media are
inserted into the bin element it is constructed with, the
one that has payloaders.